Added New lock system

+++ Assets/Scripts/Lock System/KeyLocker.cs
@@ -4,11 +4,11 @@ namespace Lock_System
 {
     public class KeyLocker : Locker
     {
-        private NetworkVariable<ushort> _registratedKey = new(0, NetworkVariableReadPermission.Everyone,
+        private NetworkVariable<int> _registratedKey = new(0, NetworkVariableReadPermission.Everyone,
             NetworkVariableWritePermission.Owner);
 
         public override void Init(int userId)
-        => _registratedKey.Value = (ushort)userId;
+            => _registratedKey.Value = (ushort)userId;
 
         public override bool CanBeOpened(int value)
             => _registratedKey.Value == value;
